+++ CHANGELOG.md
@@ -3,6 +3,30 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.0.2] - yyyy-mm-dd
+### Added
+- Dependences
+    - request
+    - request-promise-native
+-type
+    - integer
+    - cep
+    - cnpj
+    - doc
+    - number
+- length
+- exists
+- inList
+- min
+
+### Removed
+-Dependencies
+    - email-validator
+    - validar-cpf
+
+---
+
 ## [0.0.1] - 2019-08-01
 ### Added
 - required
@@ -12,3 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - boolean
 - unique
 - minLength
+
+---
+
+[EOF]